Add Item Tracer filtering
