+++ src/main/kotlin/com/peasenet/mods/esp/ModEntityItemEsp.kt
@@ -24,21 +24,23 @@
 package com.peasenet.mods.esp
 
 import com.peasenet.gavui.color.Color
-import com.peasenet.gui.mod.esp.item.GuiItemEsp
+import com.peasenet.gui.mod.GuiItemEspTracerConfig
 import com.peasenet.main.GavinsModClient
 import com.peasenet.settings.SettingBuilder
 import com.peasenet.util.RenderUtils.CHUNK_RADIUS
 import com.peasenet.util.listeners.RenderListener
 import net.minecraft.client.util.math.MatrixStack
-import net.minecraft.component.DataComponentTypes
 import net.minecraft.entity.ItemEntity
-import net.minecraft.text.PlainTextContent
-import java.util.UUID
 
 /**
  * A mod that allows the player to see an esp (a box) around items.
+ *
+ * @see RenderListener
+ * @see GuiItemEspTracerConfig
+ * @see com.peasenet.util.data.ItemEntityFilter
+ *
  * @author gt3ch1
- * @version 09-06-2024
+ * @version 01-12-2025
  * @since 03-02-2023
  */
 class ModEntityItemEsp : EntityEsp<ItemEntity>(
@@ -49,19 +51,17 @@ class ModEntityItemEsp : EntityEsp<ItemEntity>(
     RenderListener {
     init {
 
-//        colorSetting.setCallback { config.itemColor = colorSetting.color }
         val menu = SettingBuilder().setTitle("gavinsmod.mod.esp.item").buildClickSetting()
         menu.setCallback {
-            GavinsModClient.minecraftClient.setScreen(GuiItemEsp())
+            client.setScreen(GuiItemEspTracerConfig(config))
         }
-//        addSetting(colorSetting)
         addSetting(menu)
     }
 
     override fun onRender(matrixStack: MatrixStack, partialTicks: Float) {/* TODO: Work on setting filters for items */
         if (config.useItemEspFilter) {
             espList = espList.filter {
-                config.itemEspFilterList.any { filter -> filter.customNameMatches(it) }
+                config.itemFilterList.any { filter -> filter.customNameMatches(it) }
             }.toMutableList()
         }
         super.onRender(matrixStack, partialTicks)
@@ -70,37 +70,4 @@ class ModEntityItemEsp : EntityEsp<ItemEntity>(
     override fun getColor(entity: ItemEntity): Color = getColor()
 
     override fun getColor(): Color = config.itemColor
-}
-
-
-class ItemEspFilter(
-    var filterString: String,
-    var filterName: String = "",
-    var enabled: Boolean,
-    val uuid: UUID = UUID.randomUUID()
-) {
-
-    constructor(filterString: String, filterName: String) : this(filterString, filterName, true)
-
-    fun customNameMatches(itemEntity: ItemEntity): Boolean {
-        if (!this.enabled) return false
-        val stack = itemEntity.stack
-        val customName = stack.get(DataComponentTypes.CUSTOM_NAME)
-        val lore = stack.get(DataComponentTypes.LORE)
-        var matches = false
-        val regex = Regex(filterString,setOf(RegexOption.DOT_MATCHES_ALL)) // TODO: Add
-        regex.options
-        if (customName != null) {
-            val content = (customName.content as PlainTextContent.Literal).string
-            matches = content.contains(filterString)
-            matches = matches || regex.containsMatchIn(content)
-        }
-        if (lore != null) {
-            matches =
-                matches || lore.lines.any { (it.content as PlainTextContent.Literal).string.contains(filterString) }
-            matches =
-                matches || lore.lines.any { regex.containsMatchIn((it.content as PlainTextContent.Literal).string) }
-        }
-        return matches
-    }
 }
